LabelBrushController: fixed scaling of brush radius
The brush neighborhood is requested in pixels, but the radius is given
in world units. To be consistent with the BrushOverlay, the neighborhood
has to be divided by the resolution of the label volume.

The current solution creates a sphere in the x-y-plane and assumes that
the resolution in x and y is the same.