Try to work aorund NullPointerException on slower machine
There seems to be a timing issue on slower machines, resulting in un-initialized `ViewerPanel` and `NullPointerException`. This is a try to work around this. Most likely, the Atlas viewer needs to be completely initialized before adding sources. `BaseView.addViewer` will probably need to block until the viewer is successfully initialized.