Replace DatasetSpec interface by DataSource interface
In order to simplify code and reduce the number of interfaces, we decided to get rid of the DatasetSpec interface, which was used to provide both viewer (volatile) and data (non-volatile, for data processing) sources.
Instead, we introduce the DataSource interface that extends bdv's Source interface and provides access to a data representation of the visualization that bdv sees through the Source interface.