render orthoslices delayed and at max 512 x something pixels this avoids intermittent stopping during navigation and general slowness in 3D viewer while BDV panels are being updated
NOTE We could not find a meaningful way in JavaFX 3D to update
bitmap textures fast or to re-use existing textures from the scene
PhongMaterial.setSelfIlluminationMap(Image) is unfortunately the only
way to get a texture into a scene and, for unknown reasons has to be
executed in the user thread which blocks everything else.  This may
become the terminal reason to eventually leave JavaFX 3D and to seek
for alternatives.