Add switch to choose between colors from segment id or fragment id
This affects only any fragment that is within a currently active segment.
The rationale is that we can distinguish fragments within active segments from other fragments based on the active segment alpha. For fragments that are not part of an active segment it is impossible to tell which segment they belong to.