validate tag when updating textures of orthoslice meshes
When the viewer transformation is changed immediately after the painting is finished, the mode switch to single-tile may occur before the rendered tile image is received. In this case, the last rendered tile in the orthoslice view would be painted black until the new frame in the single-tile mode is rendered and received. This commit fixes the described case.