calculate texCoords depending on ratio between fullscreen size and texture size
For example, if the fullscreen size is 10px (and so is the size of the texture at screen scale 1.0), and the current texture is of size 3px at screen scale 0.25, its upscaled view would correspond to 12px in full resolution. Because of that, the max texCoord needs to be set with respect to this ratio to exclude extra pixels.