do not stretch virtual mask after interpolation because of side effects
The extra 0.5 px on each side was to make sure that the interpolated shape is fully visible in the end sections and doesn't disappear or look like discontinuous stripes. However, this has undesired side effects when painting in an orthogonal view (or very close to it) because it affects adjacent sections outside the defined span.