improve thread synchronization in MaskedSource
previously, applyMask and persistCanvas used to synchronize on this until the process is completed. This was causing the application to freeze when, for example, a new mask is requested while the last one is still being committed into the canvas. Now the application will not block anymore and instead show a dialog suggesting to wait until the process is completed.