attempt to fix race condition when explicitly setting label colors
argbCache is accessed concurrently, but the accesses are not fully synchronized due to performance reasons (only write is protected). The clearCache() method changes the collection but was not protected, which sometimes caused the stream to ignore the explicitly specified color. This fix should reduce the chance of this bug, but is not guaranteed to completely solve it.