floodfill 2d bugfix: check if inside clicked object in adjacent sections
When 2D flood-filling in an orthogonal view, there is a special case implementation that runs a genuine 2D flood-fill algorithm on a number of sections around the clicked one depending on the specified depth. If the depth is thicker than 1 section, it would run the flood-filling in adjacent sections starting from the same seed point. However, if this location in the adjacent section happens to be outside the clicked object, the outside value is used in the predicate which leads to filling the entire section.