bugfix: remove overhanging low-res parent blocks from renderer tree
When the data is downsampled in the oversampling fashion, some of the lower-resolution blocks can contain overhanging chunks of data that will not be present in the higher-res blocks that are contained in this block. Due to this, lower-resolution "halos" (or parts of the label data) could be seen even when requested to render at higher resolution. This fix takes care of this by carefully removing the overhanging lower-resolution parent blocks from the render set.