Merge pull request #340 from saalfeldlab/fix-337
[BUGFIX] Mesh resources would not be freed when not being used anymore (#337). A anonymous/lambda listener was the culprit. Explicilty naming and unbinding after use solved the problem.

* Release MeshGenerator resources when disposed

Fixes #337
Before this commit, a listener would be added to the `isManagedProperty` of the `meshSettings` to update the `MeshGenerator` with appropriate global or individual `MeshSettings`. For some reason, this kept the internal JavaFX classes `NGMeshView` and `MeshView` alive, prohibiting clean-up of their resources, and, as a result, leaking memory. This commit fixes this issue by adding a `meshSettingsProperty` to the `MeshGenerator` that the caller can `unbind` to remove the listener that holds on to the `MeshManager`'s resources.

* Set manager mesh settings to null and simplify removeAll

* Revert back to unbinding mesh settings on separate thread

After @igorpisarev pointed out (and I confirmed) that unbinding is slow for a large number of selected meshes (e.g. select all on CREMI A), I had to revert the previous changes. I added a new single threaded executor to avoid creating new threads each time and inverted the call chain to avoid code duplication.