add coarsest and finest scale level sliders instead of just the highest
This is to be able to replicate the old behavior where it was possible to render where all blocks will be rendered at the specified single scale level. The rendering will still start from the lowest resolution in order to display the blocks as soon as possible, but when coarsestScaleLevel==finestScaleLevel, the rendering will always subdivide all blocks up to the desired resolution.