build scene block tree on worker thread instead of FX thread
Previously the scene block tree was computed on FX thread at the time of the scene update, which could be very slow, for example, when the coarsest scale level setting is set to 0 (meaning that it's requested to render everything at the highest resolution). Now it will be offloaded to a worker thread, and could be interrupted if the user changes the settings to less extreme values.