Remove obsolete class AxisOrder
This was never used. If it is ever to be re-introduced, will need to be thought-out better.

One of the problems is, again, the redundancy with the source transform. Should the axis order transform applied before the transform? Should it be encoded in the transform? It is unclear.

This is somewhat similar to the discussion about the meaning of `VoxelDimensions`.