fix under/over-saturation by storing two versions of texture images
if alpha channel is modified several times on the same texture image, it leads to color under- or over-saturation. To avoid this, two versions of texture image are stored: a fully opaque version, and another version with alpha channel set with respect to the opacity setting value (this one is actually displayed on the screen). If the value of the opacity setting changes, the fully opaque texture is copied from the first image into the second, and then the alpha channel is properly set in the second image with respect to the new opacity value.