3d visualize masks update mesh manager interface (#383)
Merge current master into the 3d-visualize-masks branch. Most importantly, this unifies the mesh manager interface into a single class with light wrappers to adapt to our use cases. The `MeshManager` interface has been removed and instead there is a class `AdaptiveResolutionMeshManager` that replaces the `AbstractMeshManager` (it is mostly copy & paste from that class, which was created by @igorpisarev to work with adaptive resolution meshes) and does the  bulk of the work. Instead of inheritance, light wrappers hold instances of `AdaptiveResolutionMeshManager`s and simply translate the request, e.g. updating a mesh from boolean data or activating all meshes for all selected segments. With this simplification, I was able to remove a lot of obsolete classes.

* Save WIP

* Save WIP (not working)

* Compiles now but far from finished

* Still WIP but something is showing now!

* Make block trees nullable

* Expose mesh generator state in the manager interface

* Address a few TODOs

* Enable colors

* Save WIP: got something to show for thresholding source state

* Save WIP

* Save WIP

* Bind color property to renderer in adaptive resolution renderer

* Use mesh settings in Tresholding sources

* Add some notes

* Save WIP: Update managed mesh settings

* Remove obsolete classes

* Remember state when refreshing meshes

* Add note about settings and rendererSettings

* Save WIP

separate out mesh settings from renderer settings

* Always pre-multiply mesh color

* Rename MeshManagerWithAssignmentForSegmentsKotlin to MeshManagerWithAssignmentForSegments

* Rewrite parts of MeshInfoNode

* Update use of deprecated methods

* Fix mesh settings binding issues

They would unbind to eagerly

* Only make diff-like changes when setting to current selection

* Fix formatting in MarchingCubes class

* Remove cubeSize parameter from MarchingCubes

* Remove interruption check from marching cubes

We do not need interruption at such a fine granularity

* Remove obsolete class CacheUtils

* Remove obsolete debuggin settings

Thanks @igorpisarev for catching

* Add option to disable multi-scale meshes

* Remove MeshGenerator.State.visibleProperty

This is redundant information as this is also available inMeshSettings.isVisibleProperty

* Rename Viewer3DFX.isMeshesEnabled to meshesEnabled

This is conforming with JavaFX style

* Make sure that meshes enabled/disable toggle works

* Bind enable/disable toggle of intersecting state meshes to label state meshes

* Unify mesh settings node

* Rename MeshSettings.isVisibleProperty() to visibleProperty()

* Pass isEnabled only as an argument for MeshSettingsNode.createTitledPane

This is not required for the other use cases

* Move LabelSourceStateMeshPaneNode.populateGridWithMeshSettings into MeshSettingsNode

MeshSettingsNode is the only place where this method is called.

* Re-establish mesh export button

* Simplify mesh management

* Release resources when appropriate

* Remove obsolete class InterruptibleFunctionAndCache

* Release resources in MeshManagerWithSingleMesh when removing mesh

* Remove obsolete interfaces and classes

* Remove PainteraMeshManager interface

The mesh manager does not really need an interface because controls and handling is now responsibility of the source, i.e. no need to check for certain interfaces in the handlers.

* Replace meshes rather than refresh them when render block size is changed

Generaly no need to clear the cache in that scenario

* Bind scene settings for intersecting and thresholding states as well

* Add block boundaries group to root if appropriate during construction of MeshGenerator

* Kotlinify ConnectomicsLabelState.refreshMeshes()

* Unbind from settings and properties on interrupt

* Preserve invalidate of CachedLabelBlookLookup when wrapping

* Set label block lookup relative if necessary.

* Fix max level of detail in mesh settings node

* Set levelOfDetail and finestScaleLevel in MeshSettings.setTo

* Fix formatting

* Fix kotlin compilation issue

* Optimize imports

* Remove obsolete TODOs and code

getBlockList is never an instance of Invalidate. It is sufficient to invalidate labelBlockLookup if it is Invalidate

* Remove obsolete and incorrect line of code

The line was intended to interrupt all generators whenever the 3D viewer or the manager were disabled, and replace the interrupted generators whenever they would be enabled again. The code did the exact opposite but because this update would happen in two other places, this bug would never have any consequences during use

* Rearrange methods in files more logically

* Remove redundant calls to interrupt managers

This happens in the extension function MeshGenerator<ObjectKey>.binToThis() and does not need to happen twice.

* Remove @JvmOverloads annotation that does not have any effect

* Only cancelAndUpate if explicitly specified

* Re-introduce asserts in update and updateScene

* WIP: Try to prevent freezing of UI when (de-)selecting large number of ids

* WIP: Improving UI speed when removing a large number of meshes

* Fix bug in check for manager.cancelAndUpdate

Add note that we also need to update meshes if the fragment segment assignment has changed!

* Unbind mesh settings on separate thread

* Rename SelectedIds.getActiveIds() more aptly

Indicate that it is a copy and as an array

* Add check if a segment is consistent wrt to current assignment

* Return an unmodifiable set in SelectedIds.getActiveIds()

* Rename SelectedSegments.getActiveIds() to getACtiveIdsCopyAsArray()

The name describes better what happens

* Create meshes correctly after merge

Instead of using selectedSegments.selectedIds.activeIds, use selectedSegments.selectedSegments: The mesh selection is based on the segment id, not the fragment id!

* Split bindAndUnbindService into two separate services

* Release MeshManageer.State when refreshing meshes

* Remove groups of meshes instead of single meshes

* Run state setup only in AdaptiveREsolutionMeshManager

This simplifies things

* Rewrite AdaptiveResolutionMeshManager.createMeshFor to accept state setup.

With the state setup passed to the method, all setup can happen on the `AdaptiveResolutionMeshManager`s setup thread. No need for a setup thread in wrapping mesh managers anymore, simplifies things a lot!

* Release state on single thread, as well

* Fix FragmentSegmentAssignmentOnlyLocal.isSegmentConsistent

The case of segmentId == fragmentId was not considered before. In this case, no the segmentToFragmentsMap has no entry for segmentId, and containedFragments should have exactly one entry segmentId

* Synchronize removing meshes

* Remove unnecessary safe call

* Run all updates on single thread

* Fix bug in cancelAndUpdate

This is now what it was before in the AbstractMeshManager

* Add listener only after binding to the state to avoid blocking the UI when many meshes are added simultaneously

* Always call manager.cancelAndUpdate on FX application thread

This is required but may block the UI. Come up with a fix!

* Remove unused member bindAndUnbindService

* Limit number of cancelAndUpdate calls on the FX application thread with a LatestTaskExecutor with long delay

Avoid flooding the FX application thread with thousands of calls to cancelAndUpdate() and freezing the UI for tens of seconds. Really only the latest cancelAndUpdate() call matters and it does not have to happen at high frequencey so we can add a long delay of 100 milliseconds.