use two buffers for texture images to fix threading issues
If a single image is used as the texture for the mesh view and is updated on the background thread, some strange visual bugs occur and JavaFX throws errors. Additional buffer image fixes it because only one image is used as the texture, and another one is modified by the background thread.