set opacity/shading in all texture images for consistent partial updates
If painting is initiated after changing the opacity/shading settings, the painted brush strokes will start coming in as partial updates. If only the current texture is updated on changing the settings, when painting switches to different screen scale, the texture will be inconsistent (only the painted parts of it would be updated with respect to the new settings, while the rest of the image will be displayed using the older settings). Updating all texture images when opacity/shading settings change fixes it.