don't exit shape interpolation mode if it's due to disable/enable of AllowedActions
To avoid redundant pop-ups, actions are disabled when generating a new label mask, and re-enabled after.
However, shape interpolation creates masks during interpolation mode. since we don't want to exit interpolation mode during mask generation, we need to ensure that we are not in the process of enabling/disabling actions. If we are, then don't exit, since it is expected. If we are not enabling/disabling, then we want to proceed as usualy, and exit interpolation mode