refactor: massive reoragnization
move all top-level packages that duplicated names for other packages into org.janelia.saalfeldlab

Previously, it was common practice to create an internal package name space that maps to the
package which the files should be contributed to (e.g. net.imglib2). However, with the advent of
JPMS, merging namespaces with modules no longer works, so it is best to not duplicate the module namespace.

Even better is to contribute to the desired projects