perf: support empty blocks when downsampling label multiset data
This is a memory optimization. when downsampling LMT over a cached cell image, empty blocks at higher resolution levels are still created as full-sized backing VolatileLabelMultisetArray data. This is needed for writing, but in this case, it allocates a lot of memory, and holds on for it long enough that it can cause Heap memory issues.

To this end, the downsampler cell can now return empty blocks, indicated by VolatileLabelMultisetArray data with size 0. In this case, we can just skip the block, which saves iterations as well.