fix: navigation in shape interpolation working correctly
when `super.enter()` is called for `ShapeInterpolationMode` it activates the `defaultTool` however, in `enter()` it was also switching to `shapeInterpolationTool` again, which caused it to `deactivate` shape interpolation as well. This caused the `ShapeInterpolationTool` to not be active immediately upon entering ShapeInterpolationMode. Solution is to no double-active the default tool