fix: some deadlocks with mixing synchronization methods
refactor:  use Collections.synchronizedList since it uses the returned list as the monitor, which is the same monitor as @Synchronized. Unfortunately, this means we need to manually implement observable. FxCollections.synchronizedObservableList does not use the returned list as the monitor object, unfortunately. This lead to dealock between the mutexes.

refactor:  SlicesAndInterpolants is only mutable internally, and it's API (though private) is only `List<>`